refactor(frontend): data fetching for the Map Variables forms
Move shared code for Map Variables into its own set of shared custom hooks.
