+++ frontend-v2/src/features/model/VariableRow.tsx
@@ -1,6 +1,6 @@
 // src/components/ProjectTable.tsx
 import { FC, useEffect } from "react";
-import { Control, useFieldArray, useForm } from "react-hook-form";
+import { Control } from "react-hook-form";
 import {
   TableCell,
   TableRow,
@@ -10,20 +10,17 @@ import {
   Typography,
 } from "@mui/material";
 import {
-  Variable,
-  useVariableUpdateMutation,
   CombinedModelRead,
   CompoundRead,
   ProjectRead,
   UnitRead,
   VariableRead,
 } from "../../app/backendApi";
-import useDirty from "../../hooks/useDirty";
 import { FormData } from "./Model";
 import { useSelector } from "react-redux";
 import { RootState } from "../../app/store";
 import { selectIsProjectShared } from "../login/loginSlice";
-import useEditProtocol from "./useEditProtocol";
+import { useFormData, useVariableFormState } from "./variableUtils";
 
 interface Props {
   project: ProjectRead;
@@ -55,30 +52,16 @@ const VariableRow: FC<Props> = ({
   updateLinksToPd,
   updateLagTimes,
 }) => {
-  const { fields: mappings } = useFieldArray({
-    control,
-    name: "model.mappings",
-  });
-  const { fields: derivedVariables } = useFieldArray({
-    control,
-    name: "model.derived_variables",
-  });
+  const { mappings, derivedVariables } = useFormData({ control });
+  const { addProtocol, removeProtocol, setValue, updateVariable, hasProtocol } =
+    useVariableFormState({
+      compound,
+      project,
+      timeVariable,
+      units,
+      variable,
+    });
 
-  const { handleSubmit, reset, setValue, formState, watch } = useForm<Variable>(
-    { defaultValues: variable || { id: 0, name: "" } },
-  );
-  const watchProtocolId = watch("protocol");
-  const isDirty = watchProtocolId !== variable?.protocol || formState.isDirty;
-  useDirty(isDirty);
-  const [updateVariable] = useVariableUpdateMutation();
-  const { addProtocol, removeProtocol, hasProtocol } = useEditProtocol({
-    compound,
-    project,
-    units,
-    timeVariable,
-    variable,
-    watchProtocolId,
-  });
   async function onAddProtocol() {
     const value = await addProtocol();
     if (value?.data) {
@@ -101,29 +84,6 @@ const VariableRow: FC<Props> = ({
   const isSharedWithMe = useSelector((state: RootState) =>
     selectIsProjectShared(state, project),
   );
-
-  useEffect(() => {
-    reset(variable);
-  }, [variable, reset]);
-
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      if (isDirty) {
-        handleSubmit((data) => {
-          // @ts-expect-error - lower_bound and upper_bound can be null
-          if (data.lower_bound === "") {
-            data.lower_bound = null;
-          }
-          // @ts-expect-error - lower_bound and upper_bound can be null
-          if (data.upper_bound === "") {
-            data.upper_bound = null;
-          }
-          updateVariable({ id: variable.id, variable: data });
-        })();
-      }
-    }, 1000);
-    return () => clearInterval(intervalId);
-  }, [handleSubmit, isDirty, updateVariable, variable.id]);
 
   const isPD = variable.qname.startsWith("PD");
   const linkToPD = isPD
